+++ core/monitor/MetricConstants.h
@@ -43,6 +43,12 @@ extern const std::string METRIC_PROC_PARSE_ERROR_TOTAL;
 extern const std::string METRIC_PROC_KEY_COUNT_NOT_MATCH_ERROR_TOTAL;
 extern const std::string METRIC_PROC_HISTORY_FAILURE_TOTAL;
 
+// processor filter metrics
+extern const std::string METRIC_PROC_FILTER_IN_SIZE_BYTES;
+extern const std::string METRIC_PROC_FILTER_OUT_SIZE_BYTES;
+extern const std::string METRIC_PROC_FILTER_ERROR_TOTAL;
+extern const std::string METRIC_PROC_FILTER_RECORDS_TOTAL;
+
 // processor desensitize metrics
 extern const std::string METRIC_PROC_DESENSITIZE_RECORDS_TOTAL;